restaurantReservation-login-api/src/main/java/kr/co/mentalK94/restaurantReservation/application/UserService.java
package kr.co.mentalK94.restaurantReservation.application;

import kr.co.mentalK94.restaurantReservation.domain.EmailExistedException;
import kr.co.mentalK94.restaurantReservation.domain.User;
import kr.co.mentalK94.restaurantReservation.domain.UserExistedException;
import kr.co.mentalK94.restaurantReservation.domain.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class UserService {

    private UserRepository userRepository;

    PasswordEncoder passwordEncoder;

    @Autowired
    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
        this.userRepository = userRepository;
        this.passwordEncoder = passwordEncoder;
    }

    public User authenticate(String userId, String userPassword) {

        User user = userRepository.findByUserId(userId).orElseThrow(() -> new AuthenticationWrongException());

        if (!passwordEncoder.matches(userPassword, user.getUserPassword())) {
            throw new AuthenticationWrongException();
        }

        return user;
    }
}
